% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_empresas.R
\name{download_empresa}
\alias{download_empresa}
\title{Descargar observaciones de la base de Empresas mas Grandes del País}
\usage{
download_empresa(
  url = "https://www.datos.gov.co/resource/6cat-2gcs.csv?$limit=10000"
)
}
\arguments{
\item{url}{URL del archivo CSV o recurso que se desea descargar}
}
\value{
Tibble con observaciones por empresa
}
\description{
Descarga datos desde datos.gov.co. Estándariza los nombres de columna
}
