% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_financiera.R
\name{download_financiera}
\alias{download_financiera}
\title{Descargar hasta 10.000 observaciones de inclusión financiera}
\usage{
download_financiera(
  url = "https://www.datos.gov.co/resource/ptgf-ywrb.csv?$limit=50000"
)
}
\value{
Tibble con observaciones de CUENTAS DE AHORRO y CREDITOS DE CONSUMO
}
\description{
Descarga datos desde datos.gov.co y los filtra por tipo de producto financiero. Estándariza los nombres de columna.
}
