% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_habitantes.R
\name{download_habitantes}
\alias{download_habitantes}
\title{Descarga observaciones de la base de Habitantes en situación de calle y en la calle}
\usage{
download_habitantes(url = "https://www.datos.gov.co/resource/jmky-uei2.csv")
}
\arguments{
\item{url}{URL del archivo CSV o recurso que se desea descargar}
}
\value{
Tibble con observaciones por individuo
}
\description{
Descarga datos desde datos.gov.co. Estandariza los nombres de columna
}
