% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_laboral.R
\name{download_laboral}
\alias{download_laboral}
\title{Descargar observaciones de la base de Mercado Laboral en Bogotá para los periodos 2007-2025}
\usage{
download_laboral(
  url =
    "https://datosabiertos.bogota.gov.co/dataset/199722a1-e999-422e-8214-cac75264efec/resource/51ae1760-fc6d-47d1-9c7c-669b0f3bc517/download/conjunto-mercado-laboral-bogota-03072025.xlsx"
)
}
\arguments{
\item{url}{URL del archivo CSV o recurso que se desea descargar}
}
\value{
Tibble con datos por periodo
}
\description{
Descarga datos desde datos.gov.co. Estandariza los nombes de columna
}
