% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_notas.R
\name{download_notas}
\alias{download_notas}
\title{Descargar hasta 50.000 observaciones de Saber 11 del a침o 2022}
\usage{
download_notas(
  url = "https://www.datos.gov.co/resource/kgxf-xxbe.csv?periodo=20224&$limit=50000"
)
}
\arguments{
\item{url}{URL del archivo CSV o recurso que se desea descargar}
}
\value{
Tibble con observaciones agregadas por instituci칩n educativa
}
\description{
Descarga datos desde datos.gov.co, agrupa por instituci칩n educativa, calcula promedios por m칩dulo y cuenta estudiantes.
}
