% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_turismo.R
\name{download_turismo}
\alias{download_turismo}
\title{Descargar observaciones de la base de Turismo Receptivo en el Valle del Cauca}
\usage{
download_turismo(
  url = "https://www.datos.gov.co/resource/xv5t-qgec.csv?$limit=2000"
)
}
\value{
Tibble con observaciones por individuo
}
\description{
Descarga datos desde datos.gov.co. Estandariza los nombres de columna
}
