% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_universidades.R
\name{download_universidades}
\alias{download_universidades}
\title{Descargar hasta 50.000 observaciones de Saber Pro 2022-1}
\usage{
download_universidades(
  url = "https://www.datos.gov.co/resource/u37r-hjmu.csv?periodo=20222&$limit=50000"
)
}
\arguments{
\item{url}{URL del archivo CSV o recurso que se desea descargar}
}
\value{
Tibble con observaciones por individuo
}
\description{
Descarga datos desde datos.gov.co
}
