% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vendedores.R
\name{download_vendedores}
\alias{download_vendedores}
\title{Descargar datos del Registro único de vendedores en Cartagena}
\usage{
download_vendedores(
  url = "https://www.datos.gov.co/resource/y8mp-h84j.csv?$limit=10000"
)
}
\arguments{
\item{url}{URL del archivo CSV o recurso que se desea descargar}
}
\value{
Tibble con observaciones por vendedor registrado
}
\description{
Descarga datos desde datos.gov.co. Estándariza los nombres de columna
}
