% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_finanzas}
\alias{get_finanzas}
\title{Obtener datos de inclusión financiera}
\usage{
get_finanzas()
}
\value{
Tibble con datos listos para análisis
}
\description{
Llama a la función de descarga y agrega metadatos
}
