% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_habitantes}
\alias{get_habitantes}
\title{Obtener datos de Habitantes de Calle y en la Calle}
\usage{
get_habitantes()
}
\value{
Data frame con los datos por individuo
}
\description{
Llama a la funcion de descarga y agrega metadatos
}
