% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_laboral}
\alias{get_laboral}
\title{Obtener datos de Mercado laboral en Bogota}
\usage{
get_laboral()
}
\value{
Data frame con los datos por periodo
}
\description{
Llama a la funcion de descarga y agrega metadatos
}
