% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_saber11}
\alias{get_saber11}
\title{Obtener datos de Saber 11}
\usage{
get_saber11(aplicar_transform = TRUE)
}
\arguments{
\item{aplicar_transform}{Aplicar transformación predefinida? (TRUE por defecto)}
}
\value{
Data frame con los datos
}
\description{
Carga el dataset Saber 11 incluido en el paquete
}
