% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_turismo}
\alias{get_turismo}
\title{Obtener datos de Turismo Receptivo en el Valle del Cauca}
\usage{
get_turismo()
}
\value{
Data frame con los datos por empresa
}
\description{
Llama a la funcion de descarga y agrega metadatos
}
