% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_universidades}
\alias{get_universidades}
\title{Obtener datos de resultados Saber Pro Universidades}
\usage{
get_universidades()
}
\value{
Data frame con los datos
}
\description{
Llama a la funcion de descarga y agrega metadatos
}
