% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_vendedores}
\alias{get_vendedores}
\title{Obtener datos de Vendedores registrados en Cartagena}
\usage{
get_vendedores()
}
\arguments{
\item{aplicar_transform}{Aplicar transformación predefinidad (TRUE por defecto)}
}
\value{
Data frame con los datos de los vendedores registrados
}
\description{
Llama a la función de descarga y agrega metadatos
}
