% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{indicadores_Asia}
\alias{indicadores_Asia}
\title{Obtener datos de Indicadores Financieros de Asia}
\usage{
indicadores_Asia()
}
\value{
Data frame con los datos
}
\description{
Carga el dataset Asia_final incluido en el paquete
}
