% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{indicadores_EEUU}
\alias{indicadores_EEUU}
\title{Obtener datos de Indicadores Financieros de EEUU}
\usage{
indicadores_EEUU()
}
\value{
Data frame con los datos
}
\description{
Carga el dataset EEUU_final incluido en el paquete
}
