% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{indicadores_Europa}
\alias{indicadores_Europa}
\title{Obtener datos de Indicadores Financieros de Europa}
\usage{
indicadores_Europa()
}
\value{
Data frame con los datos
}
\description{
Carga el dataset Europa_final incluido en el paquete
}
