% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{indicadores_Latinoamerica}
\alias{indicadores_Latinoamerica}
\title{Obtener datos de Indicadores Financieros de Latinoamerica}
\usage{
indicadores_Latinoamerica()
}
\value{
Data frame con los datos
}
\description{
Carga el dataset Latinoamerica_final incluido en el paquete
}
