% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{saber_11}
\alias{saber_11}
\title{Base de datos Saber 11}
\format{
Un data frame con 10772 filas y 12 columnas
}
\source{
ICFES
}
\usage{
saber_11
}
\description{
Resultados de las pruebas Saber 11
}
\keyword{datasets}
