% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_finanzas}
\alias{get_finanzas}
\title{Obtener datos de inclusión financiera}
\usage{
get_finanzas()
}
\value{
Tibble con datos listos para analisis
}
\description{
Llama a la funcion de descarga y agrega metadatos
}
