% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_laboral}
\alias{get_laboral}
\title{Obtener datos de Mercado laboral en Bogotá}
\usage{
get_laboral()
}
\arguments{
\item{aplicar_transform}{Aplicar transformación predefinida? (TRUE por defecto)}
}
\value{
Data frame con los datos por periodo
}
\description{
Llama a la función de descarga y agrega metadatos
}
